+++ help.md
@@ -1,3 +1,11 @@
 # Help on `snippets`
 
 This is a work in progress 🚧
+
+Available commands are:
+
+help
+
+version
+
+generate
